@@ -10,17 +10,17 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=7.4",
         "doctrine/orm": "^2.8",
         "siganushka/doctrine-contracts": "dev-main",
-        "symfony/framework-bundle": "^5.2|^6.0",
-        "symfony/validator": "^5.2|^6.0"
+        "symfony/framework-bundle": "^5.4|^6.0",
+        "symfony/validator": "^5.4|^6.0"
     },
     "require-dev": {
         "composer/semver": "^3.0",
-        "symfony/form": "^5.2|^6.0",
-        "symfony/phpunit-bridge": "^5.2|^6.0",
-        "symfony/serializer": "^5.2|^6.0"
+        "symfony/form": "^5.4|^6.0",
+        "symfony/phpunit-bridge": "^5.4|^6.0",
+        "symfony/serializer": "^5.4|^6.0"
     },
     "autoload": {
         "psr-4": {
