@@ -6,13 +6,22 @@ namespace Siganushka\GenericBundle\Event;
 
 class PublicFileDataEvent extends AbstractFileEvent
 {
-    protected $data = [];
+    /**
+     * @var array<mixed>
+     */
+    protected array $data = [];
 
+    /**
+     * @return array<mixed>
+     */
     public function getData(): array
     {
         return $this->data;
     }
 
+    /**
+     * @param array<mixed> $data
+     */
     public function setData(array $data): self
     {
         $this->data = $data;
