@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Siganushka\GenericBundle\Utils;
 
+use Symfony\Component\Form\DataTransformerInterface;
 use Symfony\Component\Form\Extension\Core\DataTransformer\MoneyToLocalizedStringTransformer;
 
 /**
@@ -11,7 +12,7 @@ use Symfony\Component\Form\Extension\Core\DataTransformer\MoneyToLocalizedString
  */
 class CurrencyUtils
 {
-    protected $transformer;
+    protected DataTransformerInterface $transformer;
 
     public function __construct(?int $scale = 2, ?bool $grouping = true, ?int $roundingMode = \NumberFormatter::ROUND_HALFUP, ?int $divisor = 100)
     {
