@@ -13,7 +13,7 @@ class PhoneNumber extends Constraint
 {
     public const INVALID_ERROR = 'a177cf42-f1a9-4f75-9053-895b67224530';
 
-    public $message = 'This value is not a valid phone number.';
+    public string $message = 'This value is not a valid phone number.';
 
     protected static $errorNames = [
         self::INVALID_ERROR => 'INVALID_ERROR',
